# app/Http/Controllers/TypeActivitiesController.php
<?php

namespace App\Http\Controllers;

use Illuminate\Http\Request;
use App\Models\TypeActivities;

class TypeActivitiesController extends Controller
{
    /**
     * Display a listing of the resource.
     *
     * @return \Illuminate\Http\Response
     */
    public function index()
    {
        $entreprises=TypeActivities::all();
        return view('typeActivity.typeActivity', compact('entreprises'));
    }

    /**
     * Show the form for creating a new resource.
     *
     * @return \Illuminate\Http\Response
     */
    public function create()
    {
        //
    }

    /**
     * Store a newly created resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @return \Illuminate\Http\Response
     */
    public function store(Request $request)
    {
        $create_entreprise=new TypeActivities();
        $create_entreprise->name=$request->name;
        $create_entreprise->save();
        return back();
    }

    /**
     * Display the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function show($id)
    {
        //
    }

    /**
     * Show the form for editing the specified resource.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function edit($id)
    {
        $entreprises=TypeActivities::find($id);
        return view('typeActivity.typeActivity',compact('entreprises'));
    }

    /**
     * Update the specified resource in storage.
     *
     * @param  \Illuminate\Http\Request  $request
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function update(Request $request, $id)
    {
        $entreprises=TypeActivities::find($id);
        $entreprises->name=$request->name;
        $entreprises->save();
        return redirect()->route('typeActivity');
    }

    /**
     * Remove the specified resource from storage.
     *
     * @param  int  $id
     * @return \Illuminate\Http\Response
     */
    public function remove($id)
    {
        //
        $entreprises=TypeActivities::find($id);
        $entreprises->delete();
        return redirect()->route('typeActivity');
    }
    /**
     * @OA\Get(
     *      path="/type-activities",
     *      operationId="getAllEntreprises",
     *      tags={"Entreprises"},
     *      summary="Affiche Liste De Entreprises",
     *      description="Renvoit toutes Entreprises",
     *      @OA\Response(
     *          response=200,
     *          description="Successful operation",
     *          @OA\MediaType(
     *           mediaType="application/json",
     *      )
     *      ),
     *      @OA\Response(
     *          response=401,
     *          description="Unauthenticated",
     *      ),
     *      @OA\Response(
     *          response=403,
     *          description="Forbidden"
     *      ),
     * @OA\Response(
     *      response=400,
     *      description="Bad Request"
     *   ),
     * @OA\Response(
     *      response=404,
     *      description="not found"
     *   ),
     *  )
     */
    public function apiTypeActivities()
    {
       return TypeActivities::all();
    }
}